Fix misformatted code samples in the documentation
The current examples look strange since the C++-style comments result in the rest of the line being completely commented-out. This commit changes the way those parameters are documented to make them consistent with the rest of the documentation.

Note that I was unable to test this change because I wasn't able to generate the Boost documentation locally.