Use fallocate on truncate_file
On linux platform, boost interprocess shared memory uses sharder memory
files allocated in /dev/shm which is tmpfs. Upon creating a file [1],
boost tried to resize it to a requested length [2], through a ftruncate
call, whichchanges the size, but does not actually guarantee that the
memory will be allocated. As result, when accessing the memory linux
kernel will try to allocate it and if the host is under memory pressure
the allocation attempt may fail and application will receive a SIGBUS
and crash.

This change introduces poxis_fallocate call, which guarantees memory to
be allocated or otherwise the call will fail. This behavior is easier to
handle than recovering after SIGBUS. The fallocate is supported for
tmpfs, which is used for /dev/shm since linux 3.5 [3]. For older
versions, which do not support fallocate syscall, the glibc will emulate
the behavior.

[1] https://github.com/boostorg/interprocess/blob/boost-1.68.0/include/boost/interprocess/detail/managed_open_or_create_impl.hpp#L368

[2] https://github.com/boostorg/interprocess/blob/boost-1.68.0/include/boost/interprocess/detail/managed_open_or_create_impl.hpp#L403

[3] https://elixir.bootlin.com/linux/v3.5/source/mm/shmem.c#L2721